Update all manifests and add IMDT QCOM BSP v1.0.0 manifest (#3)
This PR changes all the manifests to pull the necessary
meta layers to build the BSP for the IMDT SBC8550.
Another change has been to add a name and version
to each manifest XML as this is now used by the Yocto
build to automatically generate the version information
within the image itself.

Also note that we are pulling newer versions of the qmmf-sdk
and the GStreamer plugins as this gives us access to working
AI plugins and sample GStreamer applications from Qualcomm.